+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.8] - 2022-10-01
 - Warmup-1 everyNth
 - Warmup-2 stringTimes
+- Warmup-2 frontTimes
 
 ### Changed
 - Revised versioning